requirements.txt
django
djangorestframework
markdown
django-filter

tensoflow
keras_preprocessing
selenium
pandas
numpy
instaloader